Changed longitude and latitude variables to be decimals to hold precision correctly.
Added TimepointType to StopTime (as added to to GTFS spec 2015/02/02).
Modified StopTime tests to check TimepointType values.
Validate Timepoint values in stop times.
Allow any column in CSV files to be surrounded in double quotation marks.
Allow escaping of quotation marks in CSV files.

+++ GTFS.DB.SQLite/Collections/SQLiteShapesCollection.cs
@@ -99,8 +99,8 @@ namespace GTFS.DB.SQLite.Collections
                 return new Shape()
                 {
                     Id = x.GetString(0),
-                    Latitude = x.GetDouble(1),
-                    Longitude = x.GetDouble(2),
+                    Latitude = x.GetDecimal(1),
+                    Longitude = x.GetDecimal(2),
                     Sequence = (uint)x.GetInt64(3),
                     DistanceTravelled = x.IsDBNull(4) ? null : (double?)x.GetDouble(4)
                 };
